refactor: moved network tests to aapt2
